Create hugo.yml
To build using GitHub actions.

Updated the default script a bit based on the one at https://gohugo.io/host-and-deploy/host-on-github-pages/

I hadn't realized this was simply a file I could add to the repo. I suppose I might as well click this button on the web interface now that I've actually edited the file in it.